Breakup hex-prefix to util/hex and util/nibbles
