+++ src/util/nibbles.js
@@ -1,46 +1,4 @@
 /**
- * @param {Array} dataArr
- * @returns {Buffer} - returns buffer of encoded data
- * hexPrefix
- **/
-export function addHexPrefix (key, terminator) {
-  // odd
-  if (key.length % 2) {
-    key.unshift(1)
-  } else {
-    // even
-    key.unshift(0)
-    key.unshift(0)
-  }
-
-  if (terminator) {
-    key[0] += 2
-  }
-
-  return key
-}
-
-export function removeHexPrefix (val) {
-  if (val[0] % 2) {
-    val = val.slice(1)
-  } else {
-    val = val.slice(2)
-  }
-
-  return val
-}
-
-/**
- * Determines if a key has Arnold Schwarzenegger in it.
- * @method isTerminator
- * @private
- * @param {Array} key - an hexprefixed array of nibbles
- */
-export function isTerminator (key) {
-  return key[0] > 1
-}
-
-/**
  * Converts a string OR a buffer to a nibble array.
  * @method stringToNibbles
  * @private
